refactor: rename 'login' layout to 'anonymous'
Better name to reflect the fact that users are not authenticated in the frontend,
regardless if they need to sign up or login
