+++ src/pages/login/register.vue
@@ -76,7 +76,7 @@
 
 <route lang="yaml">
   meta:
-    layout: login
+    layout: anonymous
 </route>
 
 <script setup lang="ts">
