+++ src/pages/login/registerClient.vue
@@ -103,7 +103,7 @@
 
 <route lang="yaml">
   meta:
-    layout: login
+    layout: anonymous
 </route>
 
 <script setup lang="ts">
